--- backend/models/game.model.js
@@ -1,37 +1,0 @@
-import mongoose from 'mongoose';
-
-const gameSchema = new mongoose.Schema({
-    title: {
-        type: String,
-        required: true
-    },
-    year: {
-        type: Number,
-        required: true
-    },
-    image: {
-        type: String,
-        required: true
-    },
-    genre: {
-        type: [String],
-        required: true
-    },
-    description: {
-        type: String,
-        required: true
-    },
-    sumrating: {
-        type: Number,
-        required: true
-    },
-    reviews: {
-        type: Number,
-        required: true
-    }
- })
-
- const Game = mongoose.model('Game', gameSchema);
-
- export default Game;
-
