+++ frontend/src/App.jsx
@@ -23,13 +23,5 @@ function App() {
         </>
     )
 }
-
-
-return (
-  <>
-    <ListPage />
-  </>
-  ) 
-}
  
 export default App
